requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.3
djangorestframework==3.13.1
drf-yasg==1.20.0
filelock==3.19.1
fsspec==2025.9.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
MarkupSafe==2.1.1
mpmath==1.3.0
networkx==3.4.2
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-dotenv==1.1.1
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
sympy==1.14.0
threadpoolctl==3.1.0
torch==1.13.0
triton==3.4.0
typing_extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.0.0
